moved all of app into individual files
